Fix app crash when exiting without scanning card. Fix Big Endian convertion for balance

+++ lib/tullave/tullave_worker_i.h
@@ -3,10 +3,13 @@
 #include "tullave_worker.h"
 
 #include <furi.h>
+#include <furi_hal_nfc.h>
+#include <float.h>
+
 #include <core/string.h>
 
 struct TuLlaveInfo {
-    uint32_t balance;
+    double balance;
     FuriString* card_number;
 };
